Evaluate Data Quality of AmsterdamUMCdb OMOP CDM after ETL.

This function runs the tests of the OHDSI Data Quality Dashboard. Based on: https://ohdsi.github.io/DataQualityDashboard/articles/DataQualityDashboard.html

Usage

execute_dqd_checks(custom_control = TRUE, check_names = c(), sql_only = FALSE)

Arguments

custom_control

Specifies whether to apply the customized control files (thresholds) for AmsterdamUMCdb Default: TRUE

check_names

Vector (c()) of data quality checks to run. Names can be found at: https://github.com/OHDSI/DataQualityDashboard/blob/main/inst/csv/OMOP_CDMv5.4_Check_Descriptions.csv Default: c()

sql_only

Specifies to output the SQL query files instead of executing them. The files will be stored in the sql folder of the dqd output folder as specified in config.yaml. Default: FALSE

Details

This function runs DataQualityDashboard::executeDqChecks() to perform the data quality checks on the AmsterdamUMCdb CDM instance. After a successful execution a json results file will be created, that can be shown in the DataQualityDashboard Shiny app using amstel::data_quality_dashboard().