Load Usagi Concepts

Reads a (partially mapped) concept list based on the concept group name. This function assumes the mapped concept list had been stored in the data/mappings folder.

Usage

load_usagi_concepts(concept_group)

Arguments

concept_group

concept group name that will be used to read the file

Value

tibble containing the concepts belonging to this group

Examples

providers <- load_usagi_concepts("providers")
providers
#> # A tibble: 36 × 17
#>    sourceCode       sourceName sourceFrequency sourceAutoAssignedCo…¹ matchScore
#>    <chr>            <chr>                <dbl> <lgl>                       <dbl>
#>  1 NEU_Neurologie   NEU_Neuro…           73175 NA                           0.37
#>  2 NCH_Neurochirur… NCH_Neuro…            6304 NA                           0.46
#>  3 Technisch        Technical            29411 NA                           0.71
#>  4 DIET_Diëtetiek   DIET_Diet…             186 NA                           0.2 
#>  5 KNF_Klinische n… KNF_Clini…              98 NA                           0.5 
#>  6 ICV_Coassistent  ICV_Coass…           85081 NA                           0.22
#>  7 OOG_Oogheelkunde EYE_Ophth…              22 NA                           0.48
#>  8 MMI_Medische mi… MMI_Medic…           99140 NA                           0.41
#>  9 Consulent        Consultant            9938 NA                           0.81
#> 10 NEF_Nefrologie   NEF_Nephr…              70 NA                           0.4 
#> # ℹ 26 more rows
#> # ℹ abbreviated name: ¹​sourceAutoAssignedConceptIds
#> # ℹ 12 more variables: mappingStatus <chr>, equivalence <chr>,
#> #   statusSetBy <chr>, statusSetOn <dbl>, conceptId <dbl>, conceptName <chr>,
#> #   domainId <chr>, mappingType <chr>, comment <chr>, createdBy <chr>,
#> #   createdOn <dbl>, assignedReviewer <lgl>