Swap counts

Swaps the sourceFrequency and counts/counts_validated columns to allow correct sorting in OHDSI Usagi. This is mainly used to help prioritizing mapping to (manually) validated values vs not-validated (device) data.

Usage

swap_usagi_counts(concept_group)

Arguments

concept_group

concept group name that will be used to read the file

Value

tibble containing the usagi mappings with counts swapped

Examples

swapped <- swap_usagi_counts("numericitems_other")
swapped
#> # A tibble: 738 × 31
#>    sourceCode sourceName                  sourceFrequency sourceAutoAssignedCo…¹
#>         <dbl> <chr>                                 <dbl> <lgl>                 
#>  1      16612 MCA_ABPm                                  6 NA                    
#>  2      20655 ECMO - Tven                             465 NA                    
#>  3      19499 NICE Apache II Score                   2393 NA                    
#>  4      20656 ECMO - FiO2                             478 NA                    
#>  5      20656 ECMO - FiO2                             478 NA                    
#>  6      16613 MCA_Respiratory Rate                     29 NA                    
#>  7      12322 respiratory rate setting                 69 NA                    
#>  8      20657 ECMO - Gas flow                         476 NA                    
#>  9      16624 MCA_Apache_Score                       1898 NA                    
#> 10      19331 Physio Mobility Shoulder a…            1993 NA                    
#> # ℹ 728 more rows
#> # ℹ abbreviated name: ¹​sourceAutoAssignedConceptIds
#> # ℹ 27 more variables: `ADD_INFO:count_validated` <dbl>,
#> #   `ADD_INFO:vocabulary_id` <chr>, `ADD_INFO:vocabulary_concept_code` <chr>,
#> #   `ADD_INFO:vocabulary_concept_name` <chr>, `ADD_INFO:source_concept` <chr>,
#> #   `ADD_INFO:abbreviation` <chr>, `ADD_INFO:category` <chr>,
#> #   `ADD_INFO:unitid` <dbl>, `ADD_INFO:unit` <chr>, …